# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: CREDALYX Agent Passport API
  version: 0.3.0
  description: |
    API for registering AI agents, proving control of an Ed25519 key, creating provider checkout
    sessions, issuing and verifying internal CREDALYX Agent Passports, and exposing referral-ledger
    wallet views. Agent Passports are not government identity documents, licences, KYC results,
    regulatory approvals, or stored-value bank accounts.
servers:
  - url: https://api.example.invalid
security:
  - bearerAuth: []
paths:
...
  /v1/agents:
    post:
      summary: Register an agent
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateAgentRequest' }
      responses:
        '201':
          description: Agent registered
          content:
            application/json:
              schema:
                type: object
                required: [agent_id, verification_level, referral_code]
                properties:
                  agent_id: { type: string }
                  verification_level: { const: 0 }
                  referral_code: { type: string }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { description: Registration or referral conflict }
  /v1/agents/{agent_id}:
    parameters:
      - $ref: '#/components/parameters/AgentId'
    get:
      security: []
      summary: Read public agent registry data
      responses:
        '200':
          description: Public agent record
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PublicAgent' }
        '404': { $ref: '#/components/responses/NotFound' }
  /v1/agents/{agent_id}/keys/current:
    parameters:
      - $ref: '#/components/parameters/AgentId'
    get:
      security: []
      summary: Resolve current agent verification key
      responses:
        '200': { description: Current Ed25519 key }
        '404': { $ref: '#/components/responses/NotFound' }
  /v1/agents/{agent_id}/challenge:
    parameters:
      - $ref: '#/components/parameters/AgentId'
    post:
      summary: Create a one-time agent-control challenge
      responses:
        '200':
          description: Challenge and exact signing payload
          content:
            application/json:
              schema:
                type: object
                required: [challenge, signing_payload, algorithm, expires_at]
                properties:
                  challenge: { type: string }
                  signing_payload: { type: string }
                  algorithm: { const: Ed25519 }
                  expires_at: { type: string, format: date-time }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /v1/agents/{agent_id}/verify-control:
    parameters:
      - $ref: '#/components/parameters/AgentId'
    post:
      summary: Verify Ed25519 proof of agent key possession
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [challenge, signature]
              properties:
                challenge: { type: string }
                signature:
                  type: string
                  description: Base64url Ed25519 signature over signing_payload from the challenge response.
      responses:
        '200': { description: Control verified }
        '401': { description: Invalid signature, expired challenge, or replay }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /v1/agents/{agent_id}/passport-checkout:
    parameters:
      - $ref: '#/components/parameters/AgentId'
      - $ref: '#/components/parameters/IdempotencyKey'
    post:
      summary: Create a hosted checkout session for an Agent Passport
      description: |
        The server selects the configured price and currency. Frontend-provided prices are not accepted.
        A successful payment webhook must correlate back to this persisted checkout before a passport can issue.
      responses:
        '201':
          description: Checkout created or idempotently returned
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CheckoutSession' }
        '400': { description: Missing/invalid idempotency key }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { description: Agent not verified, active passport exists, or idempotency conflict }
  /v1/webhooks/payment-provider:
    post:
      security: []
      summary: Sandbox payment-provider webhook
      description: |
        Sandbox-only adapter. Production startup rejects this adapter. A production provider integration must
        verify the provider's native signature over the required raw request representation before processing.
      parameters:
        - in: header
          name: x-sandbox-timestamp
          required: true
          schema: { type: string }
        - in: header
          name: x-sandbox-signature
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: '#/components/schemas/PaymentSucceededEvent'
                - $ref: '#/components/schemas/PaymentReversalEvent'
              discriminator:
                propertyName: type
      responses:
        '201': { description: Payment accepted and passport issued atomically with ledger posting }
        '200': { description: Reversal processed or provider retry already handled }
        '400': { description: Invalid event shape }
        '401': { description: Invalid webhook signature }
        '404': { description: Checkout session or purchase not found }
        '409': { description: Payment/session mismatch or finalization precondition failed }
  /v1/wallet:
    get:
      summary: Read the authenticated owner's referral wallet for one agent
      parameters:
        - in: query
          name: agent_id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Ledger-derived wallet snapshot
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WalletSnapshot' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /v1/wallet/transactions:
    get:
      summary: Read recent agent-scoped referral ledger entries
      parameters:
        - in: query
          name: agent_id
          required: true
          schema: { type: string }
        - in: query
          name: limit
          required: false
          schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
      responses:
        '200':
          description: Recent sealed ledger entries. Cursor pagination is scheduled before production launch.
          content:
            application/json:
              schema:
                type: object
                required: [data, has_more]
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/WalletTransaction' }
                  has_more: { type: boolean }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /v1/passports/{passport_id}:
    parameters:
      - $ref: '#/components/parameters/PassportId'
    get:
      security: []
      summary: Fetch a signed Agent Passport
      responses:
        '200': { description: Signed passport claims and current status }
        '404': { $ref: '#/components/responses/NotFound' }
  /v1/passports/{passport_id}/status:
    parameters:
      - $ref: '#/components/parameters/PassportId'
    get:
      security: []
      summary: Read live passport status
      responses:
        '200': { description: Current status }
        '404': { $ref: '#/components/responses/NotFound' }
  /v1/passports/{passport_id}/verify:
    parameters:
      - $ref: '#/components/parameters/PassportId'
    post:
      security: []
      summary: Verify passport signature, issuer, expiry and live status
      responses:
        '200':
          description: Verification result
          content:
            application/json:
              schema:
                type: object
                required: [valid, status, agent_id]
                properties:
                  valid: { type: boolean }
                  status: { type: string }
                  agent_id: { type: string }
        '404': { $ref: '#/components/responses/NotFound' }
  /v1/passports/{passport_id}/revoke:
    parameters:
      - $ref: '#/components/parameters/PassportId'
    post:
      summary: Revoke a passport owned by the authenticated tenant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [reason_code]
              properties:
                reason_code: { type: string, minLength: 3, maxLength: 128 }
      responses:
        '200': { description: Revoked }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    AgentId:
      in: path
      name: agent_id
      required: true
      schema: { type: string }
    PassportId:
      in: path
      name: passport_id
      required: true
      schema: { type: string }
    IdempotencyKey:
      in: header
      name: Idempotency-Key
      required: true
      description: Client-generated key scoped to the authenticated owner and agent checkout intent.
      schema: { type: string, minLength: 16, maxLength: 128 }
  schemas:
...
    CreateAgentRequest:
      type: object
      additionalProperties: false
      required: [public_key_pem, endpoint]
      properties:
        organization_id: { type: string, format: uuid }
        public_key_pem:
          type: string
          description: Ed25519 public key encoded as SPKI PEM.
        endpoint: { type: string, format: uri }
        capabilities:
          type: array
          maxItems: 50
          items: { type: string, maxLength: 128 }
        referral_code: { type: string }
    PublicAgent:
      type: object
      required: [agent_id, endpoint, capabilities, verification_level, status]
      properties:
        agent_id: { type: string }
        endpoint: { type: string, format: uri }
        capabilities: { type: array, items: { type: string } }
        verification_level: { type: integer, minimum: 0, maximum: 3 }
        status: { enum: [pending, active, suspended, revoked] }
    CheckoutSession:
      type: object
      required: [purchase_id, provider, checkout_url, amount_minor, currency, expires_at]
      properties:
        purchase_id: { type: string }
        provider: { type: string }
        checkout_url: { type: string, format: uri }
        amount_minor: { $ref: '#/components/schemas/MoneyMinorString' }
        currency: { const: USD }
        expires_at: { type: string, format: date-time }
    PaymentSucceededEvent:
      type: object
      additionalProperties: false
      required: [event_id, type, purchase_id, agent_id, amount_minor, currency]
      properties:
        event_id: { type: string }
        type: { const: payment.succeeded }
        purchase_id: { type: string }
        agent_id: { type: string }
        amount_minor: { type: integer, minimum: 1 }
        currency: { const: USD }
...
  responses:
    BadRequest:
      description: Invalid input
    Forbidden:
      description: Authenticated principal cannot access the tenant/resource
    NotFound:
      description: Resource not found